requirements.txt
python-binance==1.0.19
pandas==2.2.3
numpy==2.1.1
matplotlib==3.9.2
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.35.0
pyarrow==17.0.0